(PC-27864) refactor(tuto): passation
