+++ src/features/tutorial/components/onboarding/welcomeBackground.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+export const WELCOME_BACKGROUND_SOURCE = require('../../../../../public/images/WelcomeBackground.jpg')
