--- src/features/tutorial/components/welcomeBackground.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-export const WELCOME_BACKGROUND_SOURCE = require('../../../../public/images/WelcomeBackground.jpg')
