+++ src/features/tutorial/pages/onboarding/OnboardingAgeInformation.tsx
@@ -9,7 +9,7 @@ import {
   UseNavigationType,
 } from 'features/navigation/RootNavigator/types'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
-import { OnboardingTimeline } from 'features/tutorial/components/OnboardingTimeline'
+import { OnboardingTimeline } from 'features/tutorial/components/onboarding/OnboardingTimeline'
 import { TutorialTypes } from 'features/tutorial/enums'
 import { TutorialPage } from 'features/tutorial/pages/TutorialPage'
 import { analytics } from 'libs/analytics'
